project/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain-ibm==0.1.0
pinecone-client==2.2.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
requests==2.31.0
python-multipart==0.0.6